Add `GitLike` (#67)
- This adds a bunch of parameterization to the Git interface:
- `Git` is now parameterized by the type of the type of the current dir
path. This means we don't have to clone a path to execute a Git command
in another directory as often.
- A new `GitLike` trait represents types that can be dereferenced into a
`&Git`. This makes it possible to parameterize types like `GitWorktree`
by the type of git wrapper they contain, making it possible to write
methods that are only available for `AppGit`.