Remove footgun with `add.branch_replacements`
If you defined `add.branch_replacements` in your `config.toml`, but the
result of applying the replacements included a `/`, `git-prole` would
naively use that as the basename of a directory, causing problems (like
the parent directory not existing).

Now, the last component of the result is used, to match the behavior
when `add.branch_replacements` is not set.